% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_bargraph_frequency.R
\name{construct_bargraph_frequency}
\alias{construct_bargraph_frequency}
\title{Plot a bargraph of frequencies}
\usage{
construct_bargraph_frequency(df, x, y_points, x_scale = 0.1, y_scale = 0.05)
}
\arguments{
\item{df}{The dataset to access data from}

\item{x}{The variable to display the distribution of}

\item{y_points}{The number of ticks to place on the y axis}

\item{x_scale}{Handles the rounding of the bin intervals. For example, 0.1 would set the bin groupings so they occur between two one decimal place numbers}

\item{y_scale}{Handles the rounding of the y axis. Set to 0.05 (or 5\%) by default}
}
\value{
A ggplot2 object
}
\description{
This function provides a template for producing bargraphs in the Economic Farm Survey that
can be used to display histogram data on a bargraph (expressed as a percentage). Also
includes custom tooltip text that allows this percentage to be shown in HTML output
}
